requirements.txt
flask
sqlalchemy
pytest
gunicorn
psycopg2
python-dotenv